refactor: move test drift + run/baseline result types to homeboy-extension-contract
Continues the test subsystem breakdown with a batch of pure result types:
- Drift closure: ProductionChange, ChangeType, DriftedTest, DriftReport
- RawTestOutput (test/run.rs)
- AutoFixDriftOutput (test/workflow.rs)
- TestBaselineComparison (test/baseline.rs; unblocked now that its only field
  dep TestCounts is already in the contract crate)

The drift detection functions, baseline fingerprinting, and run/workflow
behavior stay in core. Moved the is_zero serde helper the baseline type needs.
Core re-exports all types so crate::extension::test paths stay stable.

These clear more of the TestRunWorkflowResult / MainTestWorkflowResult /
TestCommandOutput field graph toward moving the aggregate test result cluster.
