+++ crates/homeboy-extension-contract/src/lib.rs
@@ -23,6 +23,7 @@ pub mod capability;
 pub mod test_analysis;
 pub mod test_parsing;
 pub mod test_result;
+pub mod test_workflow;
 pub mod trace_parsing;
 pub use bench_artifact::{BenchArtifact, BenchArtifactViewer, BenchPreviewLifecycleMetadata};
 pub use bench_diagnostics::{
@@ -46,6 +47,10 @@ pub use test_analysis::{
 };
 pub use test_parsing::{CoverageOutput, TestFailureSummaryItem, TestSummaryOutput, UncoveredFile};
 pub use test_result::{TestCounts, TestScopeOutput};
+pub use test_workflow::{
+    AutoFixDriftOutput, ChangeType, DriftReport, DriftedTest, ProductionChange, RawTestOutput,
+    TestBaselineComparison,
+};
 pub use trace_parsing::{
     TraceArtifact, TraceAssertion, TraceAssertionStatus, TraceCanonicalCheck,
     TraceComponentsProvenance, TraceDependencyProvenance, TraceEvent, TraceEvidenceMetadata,
